Commit important code related to first letter typing
I saw that I had uncommitted code in this file yesterday, and figured it was unimportant, so I ran 'git checkout' and got rid of it. This morning I realized I shouldn't have done that. Fortunately saw that my code editor hadn't reloaded the file after 'git checkout' -- so I put it back and am now committing. I'll update it on the server too so that the first letter typing works as it should for anybody trying it.